Update annotate.info
he name element is used to display the name of the module on the Modules configuration page. The description element describes the module and is also displayed on the Modules configuration page. The package element defines which package or group the module is associated with. On the Modules configuration page, modules are grouped and displayed by package. The Core field defines the version of Drupal the module was written for. The php element defines what version of PHP is required by the module. And, the files element is an array of the names of the files that are associated with the module. In the case of the annotation module, the files associated with this module are the annotate.module and annotate.install files.