Create annotate.module
First, note the comment style. We begin with /**, and on each succeeding line, we use a single asterisk indented with one space ( *) and */ on a line by itself to end a comment. The @file token denotes that what follows on the next line is a description of what this file does. This one-line description is used so that api.module (see http://drupal.org/project/api),Drupal’sautomated documentation extractor and formatter, can find out what this file does.