Update annotate.module
update the .module file to include a check to see whether the person viewing a node is the author of that node. If the person is not the author, then we want to hide the annotation from that user. I’ll take a simple approach of using hook_node_load(), the hook that is called when a node is being loaded. In the hook_node_load() function, I’ll check to see whether the person viewing the node is the author.